Display revisions within the FileAction view
This avoids the need to use Redux
