+++ packages/website/src/pages/ContentPage/DocPage.tsx
@@ -1,16 +1,11 @@
 import { InlineLoading } from 'carbon-components-react';
-import dayjs from 'dayjs';
-import { Stack } from 'office-ui-fabric-react';
-import styles from '../FileAction.module.scss';
-import Avatar from 'react-avatar';
 import React, { useCallback, useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
 import { useHistory } from 'react-router-dom';
 import { useManagedRenderStack } from '../../context/RenderStack';
 import { history, DriveFile, parseDriveLink } from '../../utils';
 import { fromHTML } from '../../utils/docHeaders';
 import { setHeaders } from '../../reduxSlices/headers';
-import { selectRevisions, disableRevisions } from '../../reduxSlices/files';
 
 export interface IDocPageProps {
   file: DriveFile;
@@ -100,11 +95,8 @@ function DocPage({ file, renderStackOffset = 0 }: IDocPageProps) {
   const [docContent, setDocContent] = useState('');
   const [isLoading, setIsLoading] = useState(true);
   const revs: Array<gapi.client.drive.Revision> = []
-  const [revisions, setRevisions] = useState(revs);
-  const [viewRevision, setViewRevision] = useState(0);
   const history = useHistory();
   const dispatch = useDispatch();
-  const revisionsEnabled = useSelector(selectRevisions);
 
   useManagedRenderStack({
     depth: renderStackOffset,
@@ -166,26 +158,6 @@ function DocPage({ file, renderStackOffset = 0 }: IDocPageProps) {
     };
   }, [file.id, dispatch, setDocWithPlainText, setDocWithRichContent]);
 
-  useEffect(() => {
-    const fields = "revisions(id, modifiedTime, lastModifyingUser, exportLinks)"
-    async function loadRevisions() {
-      try {
-        const resp = await gapi.client.drive.revisions.list({fileId: file.id!, fields})
-        setRevisions(resp.result.revisions!.reverse());
-      } catch(e) {
-        console.error('DocPage files.revisions', file.id, e);
-      }
-    }
-
-    if (revisionsEnabled) {
-      loadRevisions();
-    } else {
-      setRevisions([]);
-    }
-
-    return function(){ dispatch(disableRevisions()) }
-  }, [file.id, revisionsEnabled]);
-
   const handleDocContentClick = useCallback(
     (ev: React.MouseEvent) => {
       if (isModifiedEvent(ev)) {
@@ -202,35 +174,6 @@ function DocPage({ file, renderStackOffset = 0 }: IDocPageProps) {
 
   return (
     <div style={{ maxWidth: '50rem' }}>
-      {(revisions.length > 0) && (
-        <div className="revisions">
-          {revisions.map((revision) => {
-            const htmlLink = (revision.exportLinks ?? {})["text/html"];
-            return (<Stack
-              key={revision.id}
-              verticalAlign="center"
-              horizontal
-              tokens={{ childrenGap: 8 }}
-              className={styles.note}
-            >
-              <a href={htmlLink}>
-                {dayjs(revision.modifiedTime).fromNow()}
-              </a>
-              <Avatar
-                name={revision.lastModifyingUser?.displayName}
-                src={revision.lastModifyingUser?.photoLink}
-                size="20"
-                round
-              />
-              <span>
-                {revision.lastModifyingUser?.displayName}
-              </span>
-            </Stack>
-            )
-          })
-        }
-        </div>
-      )}
       {isLoading && <InlineLoading description="Loading document content..." />}
       {!isLoading && (
         <div
