+++ packages/website/src/pages/FileAction.tsx
@@ -6,9 +6,8 @@ import {
   IContextualMenuItem,
   Stack,
 } from 'office-ui-fabric-react';
-import React, { useMemo } from 'react';
+import React, { useMemo, useEffect, useState } from 'react';
 import Avatar from 'react-avatar';
-import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { DriveIcon } from '../components';
 import { getConfig } from '../config';
@@ -22,13 +21,18 @@ import { showMoveFile } from './FileAction.moveFile';
 import { showRenameFile } from './FileAction.renameFile';
 import { showTrashFile } from './FileAction.trashFile';
 import responsiveStyle from '../layout/responsive.module.scss';
-import { toggleRevisions } from '../reduxSlices/files';
 
 function FileAction() {
   // Support we have a folder, containing a shortcut to a README document,
   // the rInner is README and the rOuter is the folder.
   const { inMost: rInner, outMost: rOuter } = useRender();
-  const dispatch = useDispatch();
+  const [revisionsEnabled, setRevisionsEnabled] = useState(false);
+  function toggleRevisions() {
+    console.info("toggling revisions");
+    setRevisionsEnabled(!revisionsEnabled)
+  }
+  const revs: Array<gapi.client.drive.Revision> = []
+  const [revisions, setRevisions] = useState(revs);
 
   const history = useHistory();
 
@@ -44,7 +48,8 @@ function FileAction() {
       commands.push({
         key: 'modify_user',
         onClick: () => {
-          dispatch(toggleRevisions());
+          console.info("click toggle revisions");
+          toggleRevisions();
         },
         text: (
           <Stack
@@ -218,11 +223,60 @@ function FileAction() {
     return commands;
   }, [rInner?.file, rOuter?.file, outerFolder.file, history]);
 
+  useEffect(() => {
+    const fields = "revisions(id, modifiedTime, lastModifyingUser, exportLinks)"
+    async function loadRevisions() {
+      try {
+        const resp = await gapi.client.drive.revisions.list({fileId: rInner?.file.id!, fields})
+        setRevisions(resp.result.revisions!.reverse());
+      } catch(e) {
+        console.error('DocPage files.revisions', rInner?.file.id, e);
+      }
+    }
+
+    if (revisionsEnabled) {
+      loadRevisions();
+    }
+  }, [rInner?.file, revisionsEnabled]);
+
   if (commandBarItems.length === 0 && commandBarOverflowItems.length === 0) {
     return null;
   }
 
-  return <CommandBar items={commandBarItems} overflowItems={commandBarOverflowItems} />;
+  return (
+    <div>
+      <CommandBar items={commandBarItems} overflowItems={commandBarOverflowItems} />
+      {revisionsEnabled && (
+        <div className="revisions">
+          {revisions.map((revision) => {
+            const htmlLink = (revision.exportLinks ?? {})["text/html"];
+            return (<Stack
+              key={revision.id}
+              verticalAlign="center"
+              horizontal
+              tokens={{ childrenGap: 8 }}
+              className={styles.note}
+            >
+              <a href={htmlLink}>
+                {dayjs(revision.modifiedTime).fromNow()}
+              </a>
+              <Avatar
+                name={revision.lastModifyingUser?.displayName}
+                src={revision.lastModifyingUser?.photoLink}
+                size="20"
+                round
+              />
+              <span>
+                {revision.lastModifyingUser?.displayName}
+              </span>
+            </Stack>
+            )
+          })
+        }
+        </div>
+      )}
+    </div>
+  );
 }
 
 export default React.memo(FileAction);
